Move form to a separate component
